feat: composer-unused.php file, part of QA config
composer-unused is going to be used for PHP QA purposes, however the
tool in certain situations will report false positives. As a way to
ignore false positives the tool can be instructed via a
composer-unused.php file. This commit included a baseline way to
exclude such false positives as experienced on the demo-symfony-app
example project.

Going forward the generated file could be further enhanced to include
even more packages by default, or provide means via the repo.yml to
excend if it it's deemed necessary.

Relates to #7676