feat: script to merge repo and schema defaults
As the JSON Schema can contain, and it does, default values, but there
is no mechanism by default to use these values an external script was
necessary.

The script uses an additional python package; jsonschema-default to
build a dictionary with the defaults of the schema, and a utility
function from within the pydantic package in order to recursively merge
the nested dictionaries.