fix: workaround path field set before validation
While the path field is not in any way necessary in the schema, the
object that's passed to the validator (`repo`) is modified beforehand to
include the path property which causes validation to fail.

This is a temporary workaround and should be later fixed by rewriting
the order of operations within the playbooks, or just use a separate
variable across playbooks for the path without modifying the repo
variable.