feat: composer-unused repo.yaml support
Create a new entry to allow further customization for composer-unused,
especially around the types of packages that should be ignored by
default, while at the same time allowing for project specific overrides.

Switched from NamedFilter to PatternFilter as it's a more flexible
option, but more important it was easier this way to explicitly filter
out all the platform extensions (ext-*) for which I couldn't find an
exhaustive list. `composer show -p` only lists those installed on the
system.