fix: use repo_path instead of changing repo var
The `repo` variable contains the data loaded from the project specific
repo.yaml file, and `path` does not need to be part of that schema as
it's a variable specific to the playbook runtime.

The change that introduced this element to the schema was just a quick
workaround to test the playbook without needing to rewrite all the
playbooks at that time.