fix!: set hosts: all across all the playbooks
This change is necessary because the playbook must contain the hosts
directive, and hosts all allows us to reuse the playbooks when it's
imported with a list of hosts specified on the commandline/invocation.

This is a workaround in the end because the suggested approach is to
move all tasks outside of the playbooks and import those as groups
instead of importing playbooks.