feat: create a playbook-all.yaml
Runs in bulk operations across all the repositories specified within the
inventory file (importing all the other playbooks).

Based on discussion this playbook was meant to be called playbook-apply,
but since that playbook is already in use and getting changes in another
branch at the moment, have create a separate playbook for ilustration
purposes.