fix: use a handrolled schema defaults extractor
The previous library which was used to fill in the defaults for parsed
repo.yaml file would do much more than what the documentation hinted at.
Instead of only looking for the fields which had a default specified, it
would also generate on it's own data based on the constraints defined
within the jsonschema. For example random letter strings, or it would
piick the first value of an enum, etc.