fix: keep track of default repository branch
When the changes are pushed upstream in the form of pull requests, we
need to switch back the repository locally to it's default branch, in
order for followup Ansible checkout operations to run without issues. We
switch away from the default branch during playbook execution in order
to create a separate branch which to propose in GitHub as a pull
request.