fix: changes for container/packagist repo access
During codespace startup we need to have access to the docker container
registry (the base image is built in the php-docker-base repository),
and also authenticate the user in the private packagist.

We can reuse global environment variables populated by GitHub Codespaces
(`$GITHUB_USER`), and rely on codespace configured secrets that need to
be explicitly shared with the repository (`$CR_PAT`/`$PACKAGIST_TOKEN`).