fix: remove properties from repoyaml #8022
I removed `.gitignore.j2` and removed the task to generate this file
from the playbook.
I removed the unneeded properties from `repo.yaml` and a few others.
I removed the `license` property from `repo.yaml` and specified a
default in the schema.
I added a `#` to make the sentence `the following individuals...`
a comment.
Removed `README_generated.md`.