fix: incorporate PR requested changes #8019
- Remove `todo` from `repo.yaml` and `repo.schema.yaml`. I confirm this
will not longer be generated in the README.
- In `playbook-cwd.yaml`, set `devcontainer.yaml` tasks to run for
applications, libraries, and symfony-bundles, and removed the block
from `devcontainer.yaml`.
- Moved `import_tasks: tasks/php-qa.yaml` back to `playbook-cwd.yaml`.
- Updated the `codeowners` in `repo.yaml` to align with your schema
definition, updated the `CODEOWNERS.j2` Jinja templating so it
generates the correct syntax, and added more descriptions to the
`codeowners` object in the schema, including a link to CODEOWNERS
syntax for a list of allowed patterns.