ci: improve repo-ansible workflow file #8277
Use pull_request_target to avoid leaving room for third-parties to
execute a token leak attack.

Grant the GitHub token additional permissions in order to commit changes
to the master/main branch, as well as to comment on PRs.

When pull requests are initiated comment with information about the
repo-ansible changes that will take place after a merge, instead of
attempting to find a way to push the changes directly on the branch
which was used to open the PR.