fix(ci): semantic release in projects with nodejs
The previous action would overwrite the package.json file on projects
before invoking the semantic release action. This was done in order to
install additional dependencies and configure repositoryUrl (as far as I
understand that's all it did).

Those flags can be moved to arguments of the semantic release action,
without needing to modify the repository's package.json file.