fix: disable composer-unused git push check #8978
In more recent situations, where the git hook was setup it led to
ambiguous errors about paths not being correct, where it used to work
before. As to avoid tripping users on this confuration, it has been
disabled and tracked separately as part of the 8980 epic (relating to
all things PHP QA).