feat: bundle playbook-cwd as an executable #9005
The purpose of this bundling is to simplify the way other users use the
repo-ansible project, without needing to go through the manual cloning
of this repository, ansible and its dependencies installation.

For bulk operations and development, users are still expected to clone
the repository and have ansible already installed locally.