feat: reviewdog review process container #9082
Creates a new workflow step (and container) in order to be able to use
reviewdog to check code quality of php/twig/symfony based projects.

This has an opt-in mechanism, via a new reviewdog top-level property in
the schema.

Removes configuration parts, such as yamllint due to large amounts of
false positives, as well as switching to newer/alternative tools where
the old ones are deprecated (eg twigcs-fixer -> twig-cs-fixer)