fix: remove yamllint configuration #9082
Running and checking the yamllint configuration rules https://yamllint.readthedocs.io/en/stable/rules.html
it seems that most of these conflict with autogenerated yaml files files
for symfony and GitHub action workflow specific choices.

It is a simpler choice to remove yamllint from the potential review
steps, than using a very superifical configuration that amounts for a
couple of cases that would be also catched when YAML files are used
(such as strings interpreted as truty values).