fix: version errors for new configurations #0000
When repo-ansible was applied to newly setup repositories, it would fail
to run because these new repositories didn't include a version field.
Which was a recent addition in order to enable some form of migration
across repo-ansible changes.

With this change, move the migration process before all the other tasks,
and remove line that referenced the `repo` object during the migrations
instead of the `repo_yaml` object