fix: update auto-repo-ansible permission to allow self-updates #8310
When running repo-ansible during a GitHub Actions run, changes made
to the workflow files cannot be pushed unless permitted.

In the workflows, the following snippet control permissions:

```yaml
permissions:
  contents: write # allow git commits & push
  pull-requests: write # allow comments on PR
```

This works for repo contents (`contents: write`), but writing to
workflows require `actions: write` permission. This commit adds
it to the template.

However, all repos need to manually run repo-ansible and commit
and push with more permissions, either as a user with write permissions
or a with a PAT with workflows read/write permissions.