feat: add support for configurable semantic-release #9706
This is an encompassing change that improves how our semantic-release
workflow, configuration, and playbooks work.

Currently, we use `codfish/semantic-release-action@v3` GitHub Action
with a set of options specified in the workflow file. However, for
downstream repositories that need fine-grained configuration, this
can be a bit difficult to deal with.

## Config: Use `.releaserc.cjs` file

Currently, we build a `/.releaserc.cjs` file with quite meaningful
defaults. However, in the GitHub action, it uses a custom set of
options, so the `.releaserc.cjs` file does _not_ get used.

This PR fixes this by dropping the external GitHub action altogether,
and using `npx semantic-release` directly within the action. It sets
the repo URL, so repo forks can continue to work too.

Other than the repo URL, the rest of the configuration is now read
from the `.releaserc.cjs` file.

## Release Notes: Custom Issue URL support

`semantic-release` process updates the `CHANGELOG.md` file and the
commit body with links to the issues mentioned in the commit subject.

If the downstream repository does not use GitHub Issues, this can lead
to non-existing or wrong links in the release notes and the commit
message body.

`semantic-release` supports specifying a custom issue URL format with
support for placeholder.

This PR exposes a _new_ option to override it:

`repo.github_workflows.release.issue_url_format`

When set, the release notes will be linked to this URL format, with
support for placeholders such as `{{id}}` for the issue ID in the
commit subject.

## Changed files

During the `semantic-release`, it can modify files that should be part
of the release commit.

For example, the release notes are added to the `CHANGELOG.md` file.
Additionally, it may need to include changes made to other files such
as `package.json` or the `.gitrelease` file.

This PR exposes a _new_ option to specify an array of files that should
be staged in the release commit. By default, the `CHANGELOG.md` file
is included.

`repo.github_workflows.release.commit_changed_files`

If the files are not changed, nothing will happen.

If this list contains `package.json`, the `package.json` and
`package-lock.json` files will be updated to contain the new release
being created.

This adjusts the order of the semantic-release tasks to make sure the
changes from the previous task are carried over and are considered in
the next task.

## Example

Here is an example in the `repo.yaml` file containing the new options:

```yaml
github_workflows:
  release:
    issue_url_format: https://example.net/browse/PRJCT-{{id}}
    commit_changed_files:
      - "CHANGELOG.md"
      - "package.json"
      - "package-lock.json"
      - "composer.json"
      - "composer.lock"
      - ".gitrelease"
```