fix(ci): drop cache arguments for docker build action #0000
Older versions of the GitHub cache API are deprecated and cause random,
deliberate, breakages starting with the 9th of April.

While we don't use those APIs directly, the docker action we depend
upons seem to do so.

Disable the cache arguments as a stop gap and prevent future breakages
in CI runs.